requirements.txt
openai
python-dotenv
pdfplumber
tiktoken
faiss-cpu
numpy
fastapi
uvicorn
python-multipart
pytesseract
pillow
cachetools